LPS-74724 Create processor package and move it there
