+++ modules/apps/adaptive-media/adaptive-media-document-library-thumbnails-test/src/testIntegration/java/com/liferay/adaptive/media/document/library/thumbnails/internal/osgi/commands/test/AMThumbnailsOSGiCommandsTest.java
@@ -405,7 +405,7 @@ public class AMThumbnailsOSGiCommandsTest {
 
 	private static final String _PROCESSOR_CLASS_NAME =
 		"com.liferay.adaptive.media.document.library.thumbnails.internal." +
-			"AMImageEntryProcessor";
+			"processor.AMImageEntryProcessor";
 
 	private static final String _THUMBNAIL_CONFIGURATION = "thumbnail";
